Feat
Created a baseline for setting proper variables for user, shell, home, uid. Made a checker if username already exists and made an incrementer for the uid so there are no overlaps, includes the nobody user. Finally added a baseline for the here doc to append to the /etc/passwd file