Feat
Added the code for adding the group in the /etc/group file and making the home directory. Also made the code to add the directory and files from /etc/skel to the home directory of the user and changed ownership of the files and directories of the home directory. Added call for asking password using passwd